requirements.txt
antlr4-python3-runtime

black
cfpq-data
networkx~=2.6.2
pre-commit
pydot
pyformlang~=1.0.1
pyparsing<=3.0.1
pytest~=7.2.1
scipy
